Modified the DB Interface to use Async methods

+++ ComplexModelBinding/Models/IInstructorRepository.cs
@@ -1,14 +1,15 @@
 ﻿using ComplexModelBinding.Data;
+using Microsoft.EntityFrameworkCore;
 
 namespace ComplexModelBinding.Models
 {
     public interface IInstructorRepository
     {
-        void SaveInstructor(Instructor instructor);
-        IEnumerable<Instructor> GetAllInstructors();
-        void DeleteInstructor (int instructorID);
-        void UpdateInstructor (Instructor instructor);
-        Instructor GetInstructor (int instructorID);
+        Task SaveInstructor(Instructor instructor);
+        Task<IEnumerable<Instructor>> GetAllInstructors();
+        Task DeleteInstructor (int instructorID);
+        Task UpdateInstructor (Instructor instructor);
+        Task<Instructor?> GetInstructor (int instructorID);
     }
 
     public class InstructorRepository : IInstructorRepository
@@ -22,32 +23,38 @@ namespace ComplexModelBinding.Models
         /// Setup DB Context
         /// </summary>
         /// <param name="_context">The Database Context</param>
-        public InstructorRepository(ApplicationDbContext _context)
+        public InstructorRepository(ApplicationDbContext context)
         {
-            _context = _context;
+            _context = context;
         }
 
         /// <summary>
         /// Removes the Instructor that has the given ID from the DB
         /// </summary>
         /// <param name="instructorID">The ID of the Instructor being removed from the DB</param>
-        public void DeleteInstructor(int instructorID)
+        public async Task DeleteInstructor(int instructorID)
         {
             // Get the requested Instructor from the DB using it's ID
-            Instructor instructor = GetInstructor(instructorID);
+            Instructor? instructor = await GetInstructor(instructorID);
             
-            // Remove that Instructor from the database 
-            _context.Instructors.Remove(instructor);
-            _context.SaveChanges();
+            // If the given instructor is not null
+            if(instructor != null)
+            {
+                // Remove that Instructor from the database 
+                _context.Instructors.Remove(instructor);
+                await _context.SaveChangesAsync();
+            }
         }
 
         /// <summary>
         /// Gets all Instructors in the DB and returns them in a collection, ordered by FullName
         /// </summary>
         /// <returns>A collection containing all Instructors in the DB, ordered by FullName</returns>
-        public IEnumerable<Instructor> GetAllInstructors()
+        public async Task<IEnumerable<Instructor>> GetAllInstructors()
         {
-            return _context.Instructors.OrderBy(currInstructor => currInstructor.FullName).ToList();
+            return await _context.Instructors
+                                 .OrderBy(currInstructor => currInstructor.FullName)
+                                 .ToListAsync();
         }
 
         /// <summary>
@@ -55,29 +62,30 @@ namespace ComplexModelBinding.Models
         /// </summary>
         /// <param name="instructorID">The ID of the Instructor being retrieved from the DB</param>
         /// <returns>The Instructor corresponding to the given ID</returns>
-        public Instructor GetInstructor(int instructorID)
+        public async Task<Instructor?> GetInstructor(int instructorID)
         {
-            return _context.Instructors.SingleOrDefault(currInstructor => currInstructor.ID == instructorID);
+            return await _context.Instructors
+                                 .SingleOrDefaultAsync(currInstructor => currInstructor.ID == instructorID);
         }
 
         /// <summary>
         /// Adds the given Instructor to the DB
         /// </summary>
         /// <param name="instructor">The Instructor being added to the DB</param>
-        public void SaveInstructor(Instructor instructor)
+        public async Task SaveInstructor(Instructor instructor)
         {
             _context.Instructors.Add(instructor); 
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
         }
 
         /// <summary>
         /// Updates the entry for the given Instructor
         /// </summary>
         /// <param name="instructor">The Instructor whose data is being updated in the DB</param>
-        public void UpdateInstructor(Instructor instructor)
+        public async Task UpdateInstructor(Instructor instructor)
         {
             _context.Add(instructor);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
         }
     }
 }
